initial phase
