+++ app/src/processing/app/Base.java
@@ -105,12 +105,9 @@ public class Base {
   PreferencesFrame preferencesFrame;
 
   // A single instance of the library manager window
-  ContributionManagerDialog libraryManagerFrame;
-  ContributionManagerDialog toolManagerFrame;
-  ContributionManagerDialog modeManagerFrame;
-  ContributionManagerDialog exampleManagerFrame;
-  ContributionManagerDialog updateManagerFrame;
+  ContributionManagerDialog contributionManagerFrame;
 
+  
   // Location for untitled items
   static File untitledFolder;
 
@@ -391,16 +388,8 @@ public class Base {
       }
     }
 
-    libraryManagerFrame =
-      new ContributionManagerDialog(ContributionType.LIBRARY);
-    toolManagerFrame =
-      new ContributionManagerDialog(ContributionType.TOOL);
-    modeManagerFrame =
-      new ContributionManagerDialog(ContributionType.MODE);
-    exampleManagerFrame =
-      new ContributionManagerDialog(ContributionType.EXAMPLES);
-    updateManagerFrame =
-      new ContributionManagerDialog(null);
+    contributionManagerFrame =
+      new ContributionManagerDialog();
 
     // Make sure ThinkDifferent has library examples too
     nextMode.rebuildLibraryList();
@@ -1394,7 +1383,7 @@ public class Base {
    * Show the library installer window.
    */
   public void handleOpenLibraryManager() {
-    libraryManagerFrame.showFrame(activeEditor);
+    contributionManagerFrame.showFrame(activeEditor);
   }
 
 
@@ -1402,7 +1391,7 @@ public class Base {
    * Show the tool installer window.
    */
   public void handleOpenToolManager() {
-    toolManagerFrame.showFrame(activeEditor);
+    contributionManagerFrame.showFrame(activeEditor);
   }
 
 
@@ -1410,7 +1399,7 @@ public class Base {
    * Show the mode installer window.
    */
   public void handleOpenModeManager() {
-    modeManagerFrame.showFrame(activeEditor);
+    contributionManagerFrame.showFrame(activeEditor);
   }
 
 
@@ -1418,12 +1407,12 @@ public class Base {
    * Show the examples installer window.
    */
   public void handleOpenExampleManager() {
-    exampleManagerFrame.showFrame(activeEditor);
+    contributionManagerFrame.showFrame(activeEditor);
   }
 
 
   public void handleShowUpdates() {
-    updateManagerFrame.showFrame(activeEditor);
+    contributionManagerFrame.showFrame(activeEditor);
   }
