+++ app/src/processing/app/contrib/ContributionPanel.java
@@ -128,7 +128,7 @@ class ContributionPanel extends JPanel {
 
     installActionListener = new ActionListener() {
       public void actionPerformed(ActionEvent e) {
-        listPanel.contribManager.status.clear();
+        listPanel.contributionTab.status.clear();
         isInstallInProgress = true;
         ((CardLayout) barButtonCardPane.getLayout()).show(barButtonCardPane, PROGRESS_BAR_CONSTRAINT);
         if (contrib instanceof AvailableContribution) {
@@ -140,7 +140,7 @@ class ContributionPanel extends JPanel {
 
     undoActionListener = new ActionListener() {
       public void actionPerformed(ActionEvent e) {
-        listPanel.contribManager.status.clear();
+        listPanel.contributionTab.status.clear();
         if (contrib instanceof LocalContribution) {
           LocalContribution installed = (LocalContribution) contrib;
           installed.setDeletionFlag(false);
@@ -157,14 +157,14 @@ class ContributionPanel extends JPanel {
               }
             }
           }
-          listPanel.contribManager.restartButton.setVisible(toBeRestarted);
+          listPanel.contributionTab.restartButton.setVisible(toBeRestarted);
         }
       }
     };
 
     removeActionListener = new ActionListener() {
       public void actionPerformed(ActionEvent arg) {
-        listPanel.contribManager.status.clear();
+        listPanel.contributionTab.status.clear();
         if (contrib.isInstalled() && contrib instanceof LocalContribution) {
           isRemoveInProgress = true;
           ((CardLayout) barButtonCardPane.getLayout()).show(barButtonCardPane, PROGRESS_BAR_CONSTRAINT);
@@ -193,12 +193,12 @@ class ContributionPanel extends JPanel {
               reorganizePaneComponents();
               setSelected(true);
 
-              ContributionManagerDialog manager = listPanel.contribManager;
+              ContributionTab contributionTab = listPanel.contributionTab;
               boolean isModeActive = false;
               if (contrib.getType() == ContributionType.MODE) {
                 ModeContribution m = (ModeContribution) contrib;
                 // TODO there's gotta be a cleaner way to do this accessor
-                for (Editor e : manager.editor.getBase().getEditors()) {
+                for (Editor e : contributionTab.editor.getBase().getEditors()) {
                 //Iterator<Editor> iter = listPanel.contribManager.editor.getBase().getEditors().iterator();
                 //while (iter.hasNext()) {
                   //Editor e = iter.next();
@@ -211,13 +211,13 @@ class ContributionPanel extends JPanel {
               if (isModeActive) {
                 updateButton.setEnabled(true);
               } else {
-                manager.restartButton.setVisible(true);
+                contributionTab.restartButton.setVisible(true);
               }
             }
           };
-          ContributionManagerDialog manager = listPanel.contribManager;
+          ContributionTab contributionTab = listPanel.contributionTab;
           LocalContribution localContrib = (LocalContribution) contrib;
-          localContrib.removeContribution(manager.editor, monitor, manager.status);
+          localContrib.removeContribution(contributionTab.editor, monitor, contributionTab.status);
         }
       }
     };
@@ -248,7 +248,7 @@ class ContributionPanel extends JPanel {
         } else {
           final String msg = contrib.getName() +
             " is not compatible with this version of Processing";
-          listPanel.contribManager.status.setErrorMessage(msg);
+          listPanel.contributionTab.status.setErrorMessage(msg);
         }
       }
     });
@@ -306,7 +306,7 @@ class ContributionPanel extends JPanel {
     updateButton.addActionListener(new ActionListener() {
 
       public void actionPerformed(ActionEvent e) {
-        listPanel.contribManager.status.clear();
+        listPanel.contributionTab.status.clear();
         isUpdateInProgress = true;
         if (contrib.getType().requiresRestart()) {
           installRemoveButton.setEnabled(false);
@@ -328,7 +328,7 @@ class ContributionPanel extends JPanel {
             public void cancel() {
               super.cancel();
               resetInstallProgressBarState();
-              listPanel.contribManager.status.setMessage("");
+              listPanel.contributionTab.status.setMessage("");
               isUpdateInProgress = false;
               installRemoveButton.setEnabled(true);
               if (contrib.isDeletionFlagged()) {
@@ -343,7 +343,7 @@ class ContributionPanel extends JPanel {
                 //Iterator<Editor> iter = listPanel.contribManager.editor.getBase().getEditors().iterator();
                 //while (iter.hasNext()) {
                 // TODO there's gotta be a cleaner way to do this accessor
-                Base base = listPanel.contribManager.editor.getBase();
+                Base base = listPanel.contributionTab.editor.getBase();
                 for (Editor e : base.getEditors()) {
                   //Editor e = iter.next();
                   if (e.getMode().equals(m.getMode())) {
@@ -355,13 +355,13 @@ class ContributionPanel extends JPanel {
               if (isModeActive) {
                 updateButton.setEnabled(true);
               } else {
-                listPanel.contribManager.restartButton.setVisible(true);
+                listPanel.contributionTab.restartButton.setVisible(true);
               }
             }
           };
           ((LocalContribution) contrib)
-            .removeContribution(listPanel.contribManager.editor,
-                                progress, listPanel.contribManager.status);
+            .removeContribution(listPanel.contributionTab.editor,
+                                progress, listPanel.contributionTab.status);
         } else {
           updateButton.setEnabled(false);
           installRemoveButton.setEnabled(false);
@@ -689,7 +689,7 @@ class ContributionPanel extends JPanel {
 
   private void installContribution(AvailableContribution info) {
     if (info.link == null) {
-      listPanel.contribManager.status.setErrorMessage(Language.interpolate("contrib.unsupported_operating_system", info.getType()));
+      listPanel.contributionTab.status.setErrorMessage(Language.interpolate("contrib.unsupported_operating_system", info.getType()));
     } else {
       installContribution(info, info.link);
     }
@@ -729,7 +729,7 @@ class ContributionPanel extends JPanel {
           installRemoveButton.setEnabled(!contrib.isUpdateFlagged());
 
           if (isError()) {
-            listPanel.contribManager.status.setErrorMessage(Language.text("contrib.download_error"));
+            listPanel.contributionTab.status.setErrorMessage(Language.text("contrib.download_error"));
           }
           ((CardLayout) barButtonCardPane.getLayout()).show(barButtonCardPane, BUTTON_CONSTRAINT);
           isInstallInProgress = false;
@@ -744,10 +744,10 @@ class ContributionPanel extends JPanel {
         }
       };
 
-      ContributionManager.downloadAndInstall(listPanel.contribManager.editor,
+      ContributionManager.downloadAndInstall(listPanel.contributionTab.editor,
                                              downloadUrl, ad,
                                              downloadProgress, installProgress,
-                                             listPanel.contribManager.status);
+                                             listPanel.contributionTab.status);
 
     } catch (MalformedURLException e) {
       Base.showWarning(Language.text("contrib.errors.install_failed"),
